implement file encoding check
