+++ tdrs-frontend/src/components/FileUpload/FileUpload.jsx
@@ -2,7 +2,6 @@ import React, { useRef, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import fileTypeChecker from 'file-type-checker'
-import languageEncoding from 'detect-file-encoding-and-language'
 
 import {
   clearError,
@@ -14,7 +13,11 @@ import {
 } from '../../actions/reports'
 import Button from '../Button'
 import createFileInputErrorState from '../../utils/createFileInputErrorState'
-import { handlePreview, getTargetClassName } from './utils'
+import {
+  handlePreview,
+  getTargetClassName,
+  tryGetUTF8EncodedFile,
+} from './utils'
 
 const INVALID_FILE_ERROR =
   'We can’t process that file format. Please provide a plain text file.'
@@ -34,41 +37,6 @@ const INVALID_EXT_ERROR = (
     </a>
   </>
 )
-
-// The package author suggests using a minimum of 500 words to determine the encoding. However, datafiles don't have
-// "words" so we're using bytes instead to determine the encoding. See: https://www.npmjs.com/package/detect-file-encoding-and-language
-const MIN_BYTES = 500
-
-/* istanbul ignore next */
-const tryGetUTF8EncodedFile = async function (fileBytes, file) {
-  // Create a small view of the file to determine the encoding.
-  const btyesView = new Uint8Array(fileBytes.slice(0, MIN_BYTES))
-  const blobView = new Blob([btyesView], { type: 'text/plain' })
-  try {
-    const fileInfo = await languageEncoding(blobView)
-    const bom = btyesView.slice(0, 3)
-    const hasBom = bom[0] === 0xef && bom[1] === 0xbb && bom[2] === 0xbf
-    if ((fileInfo && fileInfo.encoding !== 'UTF-8') || hasBom) {
-      const utf8Encoder = new TextEncoder()
-      const decoder = new TextDecoder(fileInfo.encoding)
-      const decodedString = decoder.decode(
-        hasBom ? fileBytes.slice(3) : fileBytes
-      )
-      const utf8Bytes = utf8Encoder.encode(decodedString)
-      return new File([utf8Bytes], file.name, file.options)
-    }
-    return file
-  } catch (error) {
-    // This is a last ditch fallback to ensure consistent functionality and also allows the unit tests to work in the
-    // same way they did before this change. When the unit tests (i.e. Node environment) call `languageEncoding` it
-    // expects a Buffer/string/URL object. When the browser calls `languageEncoding`, it expects a Blob/File object.
-    // There is not a convenient way or universal object to handle both cases. Thus, when the tests run the call to
-    // `languageEncoding`, it raises an exception and we return the file as is which is then dispatched as it would
-    // have been before this change.
-    console.error('Caught error while handling file encoding. Error:', error)
-    return file
-  }
-}
 
 const load = (file, section, input, dropTarget, dispatch) => {
   const filereader = new FileReader()
